*Удаление хлама

+++ src/server/scripts/Commands/cs_misc.cpp
@@ -1886,7 +1886,7 @@ public:
         LoginDatabase.Execute(stmt);
         stmt = LoginDatabase.GetPreparedStatement(LOGIN_INS_ACCOUNT_MUTE);
         stmt->setUInt32(0, accountId);
-        stmt->setUInt64(1, notSpeakTime);
+        stmt->setUInt32(1, notSpeakTime);
         stmt->setString(2, muteBy.c_str());
         stmt->setString(3, muteReasonStr.c_str());
         LoginDatabase.Execute(stmt);
